requirements.txt
opencv-python
mediapipe==0.10.21
numpy
scipy==1.11.3
twilio==8.10.0
sounddevice
vosk
python-dotenv